Move SqliteStorageDataSource to data_source.py to avoid having it declared in __init__
