+++ clock/storage/data_source/data_sources/sqlite/component/factory.py
@@ -1,8 +1,8 @@
 from sqlite3 import Connection
 
-from clock.storage.data_source.data_sources.sqlite import QuerySqliteComponent
 from clock.storage.data_source.data_sources.sqlite.component.components.chat import ChatSqliteComponent
 from clock.storage.data_source.data_sources.sqlite.component.components.message import MessageSqliteComponent
+from clock.storage.data_source.data_sources.sqlite.component.components.query import QuerySqliteComponent
 from clock.storage.data_source.data_sources.sqlite.component.components.user import UserSqliteComponent
